Software 1
    Are you having fun with this?

Did you really expect me to not approach the idea of writing about
software in any other way? Did you expect me to not be something of a
nerd about this?

    I suppose not. Tell me about software, then.

What's to say? Mom decided that, since I was showing an interest in
computers, it might be a good thing to let me use her copy of
VisualBasic 4. From there, I just kept on going.

    Well, hold on, you're skipping over a whole bunch of stuff.

I suppose so.

    You're skipping over your dad joking that, since you spent so much
    time on the computer, that he was always worried that the FBI would
    come knocking on the door one day.

Well, he was the one who got me the computers in the first place. He
bought me a copy of RedHat 6.2 on a CD at Circuit City.

    Oh, my aching bones.

I know. Every single bit of that sentence was ancient.

Still, it's largely his fault. We strung coax throughout the house in a
simple network. He bought a file server, a copy of Windows NT, and we
worked on setting up IIS together so that we could have both a file
share as well as a way of getting those files from work for him, and my
mom's house for me.

    Very kind of him. Forward thinking.

He wanted me to be an engineer. What better way to get me into the
mindset? Besides, *stuff* was his game. Our relationship was not yet
mature enoug that we could be buddies, so instead, he did what he
thought parents were supposed to do and punished, instructed, and
showered with gifts. It's just that some of those were computers.

As many gifts bounced off of me as those that stuck and proved useful.

Either way, start a kid on VisualBasic and give her access to AngelFire,
and you're bound to wound up with at least *some* kind of nerd.