Software 4
Well, then I suppose my second job in computers was in late 2005, when I
got that job at the library. That was far more comfortable.

    Or you were far more mature, perhaps.

Maybe. Either way, it was something that I was able to actually focus
on, do a good job on. There was downtime, and sometimes it got crazy.
Sometimes we'd come into the library long before it opened and blast
music while we installed or reimaged whole swaths of computers.

Sometimes we'd dick around. Nerf footballs, library cart racing. One
time Josiah locked the surplus filing cabinet we had but did not have
the key for and we had to drill out the lock. When we got it unlocked,
the first thing he did was to lock it again. We hollered and chased him
from the room as we struggled desperately to unlock the cabinet once
again.

    It was fun.

For the most part, yes. I did some development for them, too. It was my
first software job as well as my first job in computers. I did the
Atmospheric Sciences Reading Room site. I did some campus mapping. I was
enjoying it.

Enjoying it enough that, when my future in music burned down around my
ears, I was ready enough to jump on any job offer in tech that I could
manage to pull off.

    Whether or not it was something you might actually enjoy.

Yes.